Resubmitting Project 4 Phrase Hunter
Ran into a never-ending cycle of bugs, only to later realize that I don't have to declare the gameObj inside the eventlistener, as this cuts it off from the rest of the script body... leading to a mingling of instances whenever a new game is started. Iterators start skipping iterations, etc... big headache.